refactor : 테스트 클래스 위치, 네이밍 수정
